requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pytz>=2023.3
# Optional

# Google Docs integration
google-api-python-client>=2.70.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
